leetcode/236.二叉树的最近公共祖先_todo/lowestCommonAncestor_test.go
package _36_二叉树的最近公共祖先_todo

import (
	"reflect"
	"testing"
)

func Test_lowestCommonAncestor(t *testing.T) {
	type args struct {
		root *TreeNode
		p    *TreeNode
		q    *TreeNode
	}
	tests := []struct {
		name string
		args args
		want *TreeNode
	}{
		{name: `输入：root = [3,5,1,6,2,0,8,null,null,7,4], p = 5, q = 1
输出：3
解释：节点 5 和节点 1 的最近公共祖先是节点 3 。`, args: args{
			root: &TreeNode{
				Val:   0,
				Left:  nil,
				Right: nil,
			},
			p: nil,
			q: nil,
		}},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := lowestCommonAncestor01(tt.args.root, tt.args.p, tt.args.q); !reflect.DeepEqual(got, tt.want) {
				t.Errorf("lowestCommonAncestor01() = %v, want %v", got, tt.want)
			}
		})
	}
}
